+++ rs-lang/src/ts/common/Component.ts
