+++ rs-lang/src/ts/settings.ts
@@ -1,4 +1,5 @@
 export const BASE_LINK = 'http://localhost:3000';
+// export const BASE_LINK = 'https://react-learnwords-example.herokuapp.com';
 
 export enum LocalStorageKey {
   name = 'RSLang2022AppUserName',
